Move common pair code into pair_details
